+++ admin/login.php
@@ -1,6 +1,6 @@
 <?php
 session_start();
-include '../config.php';
+include '../core/config.php';
 include 'includes/functions.php';
 
 $ps = filter_var(trim($_POST['ps']), FILTER_SANITIZE_STRING);